P004: Add python solution
This is the naive solution. With digits=3 is runs pretty fast, but takes
a long time (>10s) when computing the largest palindrome that is the
multiplication of two 4-digit numbers.

Possible optimization: Adjusting variables `x` and `y` according to the
result of `x * y` might make the solution converge faster. What's is the
logic that govern these increments/decrements?